refactor(cpu): extract cpu activation and norm primitives crate
