+++ crates/bitnet-cpu-primitives/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "bitnet-cpu-primitives"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+authors.workspace = true
+keywords.workspace = true
+categories.workspace = true
+description = "CPU-side activation and normalization primitives for BitNet inference"
+
+[dependencies]
+bitnet-common = { path = "../bitnet-common", version = "0.2.1-dev" }
